Initial Commit
This is the initial commit of the Integrity Solutions Installation System (ISIS). This commit includes the following:

* `com.is2300.isis` package that contains the startup class only at this time.
* `com.is2300.test.isis` package which contains a couple of test classes.

This project relies on the Wizard API, which is available at <a href="https://github.com/Lonzak/Wizard-API.git>GitHub in the Lonzak: Wizard-API</a> repository. I simply cloned it into NetBeans and then added that project as a Library to the ISIS project.