Work has Begun on the Primary Functionality
I have created six (6) classes in the project, and three (3) packages. The packages are:

* `com.is2300.isis.generators`
* `com.is2300.isis.pages`
* `com.is2300.isis.producers`

The classes are:

* `MakeInstaller`: This is the class that contains our `public static void main(String[] args)` start-up method. 
* `MakeInstallerProducer`: This is the class that will cause the creation of the `{installer}.jar` file that will be able to be distributed.
* `InstallerGenerator`: This is the class that actually performs the creation of the installation package.
* `GetDefInstLocPage`: This is the class that displays the page to get the default installation location in the `Wizard` API.
* `GetJARPage`: This is the class that allows the user to locate and add the JAR file to be installed by the installer wizard.
* `GetLAPage`: This is the class that displays the license agreement page in the wizard.
* `GetTitlePage`: This is the class that displays the title/welcome page in the wizard.

The only one of these that are coded thus far is the `MakeInstaller` class. The article that I am using to learn how to do this has dead links for the resources and I have reached out to the article's author to see about getting a copy of the original source code. If I don't hear back from him, I'll try and figure out these classes on my own, but the article didn't even have images of the panels to go off of in order to try and recreate the classes from the article. Hopefully the author will respond to my email and be able to get me a copy of his original source code. Until then...