ISIS Make Installer Progressing
I have gotten the following `WizardPage`s completed:

- `GetJARPage`: Allows the developer to choose their JAR file visually, or type in the path to the JAR file.
- `GetLAPage`: Allows the developer to choose their license file (plain text) visually, or to type in the path to the license text file. Provides the GPL v3 plain text license from the `com.is2300.isis.contents` package by default. Contains a checkbox that needs to be selected in order to move on, stating that they wish to use the currently displayed license.
   * We could also allow the developer to provide an HTML formatted license file, as we are displaying the license in a `JEditorPane`, so it will properly display HTML formatted documents. We may want to consider this.
- `GetTitlePage`: Welcomes the developer to the Wizard and explains what they will need to do to create their own, custom installer wizard.

I've added a package, `com.is2300.isis.canvas`, which allows us to create our `WizardPage`s visually. Once we have all elements of the interface in place the way we want them, all custom initialization, pre-init, post-init, and event code entered, all we have to do is copy the `initComponents()` and event methods, and the variable declarations section to the page we are working on. This removes the need to hand-code all of the component initialization functionality by allowing the NetBeans IDE to do it for us. We can simply delete the controls that we won't be using on a new `WizardPage` and rename, if necessary, the controls that will stay. When we delete a `JButton`, for example, any custom initialization or event handling code that we created for that component will also be automatically deleted by the IDE. By using this canvas, we are saving ourselves a lot of tedious work.

I also added the new package, `com.is2300.isis.contents`, in which I placed the following files:

- `license.txt`: This is the plain-text version of the GNU General Public License version 3.
- `title.html`: This is the welcome page for the `GetTitlePage` `WizardPage` sub-class.

We will need to add other descriptive HTML files into this package as we need them. Use the CSS style:

```css
        <style>
            body {background-color: #f0fff0;
                  font-family: "Times New Roman", Times, serif;
                  font-style: normal;
                  font-size: 11px;}
        </style >
```

for the formatting the documents displayed in the Wizard. This is a nice, clean look and it fits the styling that we are going for, which is professional.