ISIS Progress Continues
Now allowing the user to select multiple, additional JAR files to add to their installer. If the JAR files are in the same folder, the user can multi-select. If the user then needs to add more JAR files from another (or other) directory(ies), they can do so and it will add the paths to the list of additional JAR folders. 

We are going to want to do this for the Libraries page as well.