Sidebar Image not Loading; All Pages Done...Time to Create Installer
I'm trying to place a custom image on the Wizard sidebar panel, but it's not working...need to look into why not.

The information gathering pages are finished, with one exception...We need to change the `JList` to a `JTextArea` on the "Additional JAR Files page, and divide the paths entered with a semicolon, as I believe the Wizard API will only store the selected value of a `JList` in the settings map.

Once we make that one change, we are ready to start copying files into a temporary folder, compiling the installer and generating the distributable JAR file. **Time to get coding!**